requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.2
upstash-redis==1.1.0

# Test suite only
rich==13.7.1
